feat: Allow configuring Yivi / Digid LOA
The login request handler can now be configured to provide a Yivi
login button, and to set a specific Digid LOA.

Since providing a LOA higher than 'laag' for Digid disallows Yivi,
we need separate login urls for Yivi / Digid. You can now:
- Set a digid service level
- Set a flag 'useYivi' to enable Yivi. This will add a second button
  for yivi.

There's a separate template for the +yivi-page. The 'useYivi'-prop
should be considered a temporary feature flag. After launch this can be
simplified to a single template.