feat: Customize navigation based on user types (#845)
The session now stores an 'identifier' as well as a user type (`person`
/ `organisation`). This type is used when constructing the navigation
for Mijn Nijmegen.

For backwards compatiblity, the 'bsn' field in session will still be
filled for 'person' users, but not for organisations. organisations
visiting the relevant pages will receive errors until we update those.

We also create a 'navigation' object, which will be filled depending on the user type.