feat: Initial 'lazy load' implementation
Loading zaken can take a long time. This release sets an agressive
timeout value on the call to get zaken from the zaakaggregator. This
timeout will often be too short.

We handle the timeout by detecting it, and showing an error to the user,
inviting them to refresh the page. Since the zaakaggregator received the
call and caches the information, a subsequent call will most likely not
timeout, showing info to the user.

Obviously this is not particularly userfriendly. This is the initial
step, we can progressively enhance what happens by:
- First adding a button/link for refreshing, perhaps adding a longer
  timeout to the next call (since the user expects it now).
- Enhancing that button using JS, refreshing the page automatically
  after a few seconds (be careful the user should expect this).
- Enhancing that by doing a JS fetch for updated info, and refreshing
  the page inline, perhaps showing a loading indicator on first launch.